# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/datos-campo": {
      "post": {
        "operationId": "DatosCampoController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatosCampoDto"
              },
              "examples": {
                "default": {
                  "summary": "Visita con mediciones dinámicas",
                  "value": {
                    "parcela_id_fk": 1,
                    "momento_id_fk": 1,
                    "observaciones": "Sin observaciones",
                    "mediciones": [
                      {
                        "variable_id": 1,
                        "valor": "10.5"
                      },
                      {
                        "variable_id": 2,
                        "valor": "Alto"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Datos de campo creados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatosCampo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear registro de datos de campo",
        "tags": [
          "datos-campo"
        ]
      },
      "get": {
        "operationId": "DatosCampoController_findAll",
        "parameters": [
          {
            "name": "momentoId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "parcelaId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [],
                    "meta": {
                      "total": 0,
                      "page": 1,
                      "limit": 10,
                      "pageCount": 0
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar datos de campo (paginado y filtrado)",
        "tags": [
          "datos-campo"
        ]
      }
    },
    "/api/v1/datos-campo/{id}": {
      "get": {
        "operationId": "DatosCampoController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatosCampo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener datos de campo por ID",
        "tags": [
          "datos-campo"
        ]
      },
      "patch": {
        "operationId": "DatosCampoController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDatosCampoDto"
              },
              "examples": {
                "reemplazoMediciones": {
                  "summary": "Actualizar observaciones y reemplazar mediciones",
                  "value": {
                    "observaciones": "Lluvia previa a la medición",
                    "mediciones": [
                      {
                        "variable_id": 1,
                        "valor": "9.8"
                      },
                      {
                        "variable_id": 3,
                        "valor": "Bajo"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatosCampo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar datos de campo",
        "tags": [
          "datos-campo"
        ]
      },
      "delete": {
        "operationId": "DatosCampoController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "deleted": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar datos de campo",
        "tags": [
          "datos-campo"
        ]
      }
    },
    "/api/v1/datos-campo/{idVisita}/upload-foto": {
      "post": {
        "operationId": "DatosCampoController_uploadFoto",
        "parameters": [
          {
            "name": "idVisita",
            "required": true,
            "in": "path",
            "description": "ID de la visita (DatosCampo)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Imagen (jpg, jpeg, png); máx. 10MB"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ejemplo de respuesta",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 1,
                    "file_name": "foto.jpg",
                    "file_path": "/uploads/visita_1/foto.jpg",
                    "mime_type": "image/jpeg",
                    "fecha_subida": "2025-11-04T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Foto registrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FotoRegistro"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sube una foto para una visita (DatosCampo)",
        "tags": [
          "datos-campo"
        ]
      }
    },
    "/api/v1/datos-campo/mediciones/por-variable/{variableId}": {
      "get": {
        "operationId": "DatosCampoController_findMedicionesByVariable",
        "parameters": [
          {
            "name": "variableId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página (1..N)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Elementos por página",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Campo para ordenar (opcional)",
            "schema": {
              "example": "id",
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Dirección de orden",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de mediciones",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [],
                    "meta": {
                      "total": 0,
                      "page": 1,
                      "limit": 10,
                      "pageCount": 0
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar mediciones por variable (paginado)",
        "tags": [
          "datos-campo"
        ]
      }
    },
    "/api/v1/datos-campo/mediciones/por-visita/{visitaId}": {
      "get": {
        "operationId": "DatosCampoController_findMedicionesByVisita",
        "parameters": [
          {
            "name": "visitaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página (1..N)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Elementos por página",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Campo para ordenar (opcional)",
            "schema": {
              "example": "id",
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Dirección de orden",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de mediciones",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [],
                    "meta": {
                      "total": 0,
                      "page": 1,
                      "limit": 10,
                      "pageCount": 0
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar mediciones por visita (paginado)",
        "tags": [
          "datos-campo"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Ensayos Agronómicos",
    "description": "Documentación de la API para el sistema de gestión de ensayos",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MedicionDto": {
        "type": "object",
        "properties": {
          "variable_id": {
            "type": "number",
            "description": "ID de la variable a medir",
            "example": 1
          },
          "valor": {
            "type": "string",
            "description": "Valor registrado",
            "example": "10.5"
          }
        },
        "required": [
          "variable_id",
          "valor"
        ]
      },
      "CreateDatosCampoDto": {
        "type": "object",
        "properties": {
          "parcela_id_fk": {
            "type": "number",
            "description": "ID de la Parcela",
            "example": 12
          },
          "momento_id_fk": {
            "type": "number",
            "description": "ID del Momento de Evaluación",
            "example": 3
          },
          "observaciones": {
            "type": "string",
            "description": "Observaciones generales de la visita"
          },
          "mediciones": {
            "description": "Lista de mediciones tomadas",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicionDto"
            }
          }
        },
        "required": [
          "parcela_id_fk",
          "momento_id_fk",
          "mediciones"
        ]
      },
      "DatosCampo": {
        "type": "object",
        "properties": {}
      },
      "UpdateDatosCampoDto": {
        "type": "object",
        "properties": {
          "observaciones": {
            "type": "string",
            "description": "Observaciones generales de la visita"
          },
          "mediciones": {
            "description": "Lista COMPLETA de mediciones. Reemplazará las existentes si se envía.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicionDto"
            }
          }
        }
      },
      "FotoRegistro": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}